Share code for S3 operations between frontend and backend.
I had multiple almost identical copies of this previously, merge them.
