+++ src/include/zenith_s3/s3_ops.h
@@ -7,23 +7,18 @@
  * Portions Copyright (c) 1996-2021, PostgreSQL Global Development Group
  * Portions Copyright (c) 1994, Regents of the University of California
  *
- * src/bin/zenith_slicedice/s3_ops.h
+ * src/include/zenith_s3/s3_ops.h
  *
  *-------------------------------------------------------------------------
  */
 #ifndef S3_OPS_H
 #define S3_OPS_H
 
-#include "lib/stringinfo.h"
-
 typedef struct
 {
 	int			numfiles;
-
 	char	  **filenames;
-
 } ListObjectsResult;
-
 
 extern ListObjectsResult *s3_ListObjects(const char *s3path);
